[test refactor] Finish refactoring tests for `winston/common`
