+++ lib/winston/common.js
@@ -60,15 +60,6 @@ exports.setLevels = function (target, past, current) {
 };
 
 //
-// ### function longestElement
-// #### @xs {Array} Array to calculate against
-// Returns the longest element in the `xs` array.
-//
-exports.longestElement = function (xs) {
-  return
-};
-
-//
 // ### function clone (obj)
 // #### @obj {Object} Object to clone.
 // Helper method for deep cloning pure JSON objects
